♻️ Extracted other RPCs to separate file
